refactor: refactoring the main entry point
